pyproject.toml
[project]
name = "api-mynest"
version = "0.1.0"
description = "API para controlar a lógica de serviços do MyNest"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.128.0",
    "pydantic==2.12.5",
    "uvicorn==0.40.0",
    "sqlalchemy==2.0.45",
    "alembic==1.17.2",
    "python-dotenv==1.2.1"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
